requirements.txt
Django==2.0
django-extensions==1.9.8
djangorestframework==3.7.3
django-cors-headers==2.1.0
dj-database-url==0.4.2
graphene-django>=2.0
gunicorn==19.7.1
ipdb==0.11
ipython==6.3.1
psycopg2==2.7.3.2
pytz==2017.3
social-auth-app-django==2.1.0
whitenoise==3.3.1
requests==2.18.4
requests-oauthlib==0.8.0
ipdb==0.11
requests==2.18.4
